Fuck webpacker, installed Vite, updated Node, updated Docker files
