+++ Dockerfile
@@ -1,17 +1,19 @@
 FROM ruby:3.2.1
 
-RUN curl -sL https://deb.nodesource.com/setup_10.x | bash - && \
+RUN curl -sL https://deb.nodesource.com/setup_19.x | bash - && \
     curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | apt-key add - && \
     echo "deb http://dl.yarnpkg.com/debian/ stable main" | tee /etc/apt/sources.list.d/yarn.list && \
     apt-get update && \
-    apt-get install -y nodejs yarn nano
+    apt-get install -y nodejs yarn nano build-essential
 
 RUN mkdir /b33rtime
 WORKDIR /b33rtime
-COPY Gemfile /b33rtime/Gemfile
-COPY Gemfile.lock /b33rtime/Gemfile.lock
+
+COPY Gemfile ./Gemfile
+COPY Gemfile.lock ./Gemfile.lock
 RUN bundle install
-COPY . /b33rtime
+RUN yarn install
+COPY . .
 
 # Add a script to be executed every time the container starts.
 COPY entrypoint.sh /usr/bin/
