+++ Gemfile
@@ -69,3 +69,4 @@ gem 'unicorn'
 gem 'fog-aws'
 gem 'fog-local'
 gem 'certified'
+gem 'vite_rails'
