+++ Procfile.dev
@@ -1,0 +1,3 @@
+
+vite: bin/vite dev
+web: bin/rails s
