+++ app/javascript/entrypoints/application.scss
@@ -1,0 +1,3 @@
+a {
+  background: #000;
+}
