+++ package.json
@@ -14,9 +14,11 @@
   "devDependencies": {
     "@vue/compiler-sfc": "^3.1.0-0",
     "deepmerge": "^4.2.2",
-    "fibers": "^4.0.2",
     "sass": "^1.26.3",
     "sass-loader": "^8.0.2",
+    "vite": "^4.0.0",
+    "vite-plugin-ruby": "^3.1.2",
+    "@vitejs/plugin-vue": "^4.0.0",
     "webpack-dev-server": "^3.10.3"
   }
 }
